test: hacking cjs tests chore: modify target and module on tsconfig configuration
