+++ jest.esm.config.ts
@@ -1,12 +1,12 @@
-import config from './jest/jest.shared.config'
+import { buildConfiguration } from './jest/jest.shared.config'
 
-config.testMatch = [
-  '**/?(*.)+(esm).+(spec|test).+(ts|tsx|js)',
-]
-config.globals = {
-  'ts-jest': {
-    useESM: true,
+export default buildConfiguration({
+  testMatch: [
+    '**/?(*.)+(esm).+(spec|test).+(ts|tsx|js)',
+  ],
+  globals: {
+    'ts-jest': {
+      useESM: true,
+    },
   },
-}
-
-export default config
+})
