+++ jest/jest.shared.config.ts
@@ -2,18 +2,17 @@ import type { Config } from '@jest/types'
 
 // see https://jestjs.io/docs/ecmascript-modules
 
-// Sync object
-const config: Config.InitialOptions = {
-  verbose: true,
-  moduleDirectories: [
-    'node_modules',
-    'src',
-  ],
-  moduleFileExtensions: ['ts', 'js'],
-  extensionsToTreatAsEsm: ['.ts'],
-  transform: {
-    '^.+\\.ts$': 'ts-jest',
-  },
+export const buildConfiguration = (configuration: Partial<Config.InitialOptions>): Config.InitialOptions => {
+  return Object.assign({}, {
+    verbose: true,
+    moduleDirectories: [
+      'node_modules',
+      'src',
+    ],
+    moduleFileExtensions: ['ts', 'js'],
+    extensionsToTreatAsEsm: ['.ts'],
+    transform: {
+      '^.+\\.ts$': 'ts-jest',
+    },
+  }, configuration)
 }
-
-export default config
