+++ package.json
@@ -15,7 +15,7 @@
 			"import": "./dist/index.mjs"
 		}
 	},
- 	"files": [
+	"files": [
 		"dist",
 		"json",
 		"lib",
