+++ src/locate.cjs.test.ts
@@ -1,10 +1,4 @@
-import { resolve } from 'path'
 import { locate } from '../dist'
+import { locateTest } from './locate.test'
 
-const cwd = process.cwd()
-
-test('mdi resolves the json collection', () => {
-  const received = locate('mdi') as string
-  const expected = resolve(cwd, 'json', 'mdi.json').replace(/\\/g, '/')
-  expect(received).toBe(expected)
-})
+locateTest(locate)
