+++ src/locate.esm.test.ts
@@ -1,10 +1,4 @@
-import { resolve } from 'path'
+import { locateTest } from './locate.test'
 import { locate } from '.'
 
-const cwd = process.cwd()
-
-test('mdi resolves the json collection', () => {
-  const received = locate('mdi') as string
-  const expected = resolve(cwd, 'json', 'mdi.json').replace(/\\/g, '/')
-  expect(received).toBe(expected)
-})
+locateTest(locate)
