+++ tsconfig.json
@@ -1,8 +1,7 @@
 {
   "compilerOptions": {
-    "baseUrl": ".",
-    "module": "esnext",
-    "target": "ESNext",
+    "target": "es2017",
+    "module": "ESNext",
     "lib": ["ESNext"],
     "esModuleInterop": true,
     "strict": true,
